Make resource titles independent of YAML ordering
Resource titles in terraform need to be unique. Previously this was
achieved by appending the index of the element within the record list
to the title. This has the affect that if the ordering of elements
in the YAML is changed (highly likely as new records should be
inserted near related records).

This replaces the previous method (of appending an index) with a hash
calculated on the uniquely identifying elements of the record:
the label (i.e. subdomain), the record type and the data. This
also helps improve the DRYness of the code by removing duplication
from the templates.