Fix Terraform backslash escaping
Terraform requires that '\' are escaped (i.e. '\\'). This adds a step
to the record modification loop that applies this to the data.

The 6 '\'s in the substitution parameter are because gsub requires
them to be doubly escaped. The first escape is to stop normal string
escaping (e.g. '\n' for newline) the second escape is to stop gsub
treating them as capture groups.