Produce terraform JSON rather than code
In general it is preferable to produce JSON rather than code. This is
because JSON is a more well supported output and we can more easily
produce it. It also means that we no longer have to split the logic of
producing terraform between template files and the Rakefile, instead
we can produce simple hashes with the correct structure and output
them directly to JSON.

In order to facilitate the change to JSON adds a test suite which
checks the consistency of low level functions (e.g. for producing
the terraform resource titles) as well as the more general functions.