Remove "@"s from GCE record names
GCE's CloudDNS expects the name of a record to be the fully resolved
domain name, rather than the label, for the record. This means that
records added to the root of the domain using BIND's "@" syntax
end up with listings like `@.example.com.` which is incorrect.

This changes the behaviour for GCE records so that those using the
"@" received the fully value of `${var.GOOGLE_DNS_NAME}` producing
the desired effect.