Try grouping records in AWS
When writing separate records out in Route53 that had the same subdomain
but a different text value, I found that it wrote all but the last
record it. I believe this is because it was overwriting the records.

Route53 stated in the console regarding TXT records (same for NS records):
```
A text record. Enter multiple values
    on separate lines. Enclose text in
    quotation marks.
Example:
    "Sample Text Entries"
    "Enclose entries in quotation marks"
```

This means we can copy the function that was written for Google Cloud
which makes the code more consistent. I have confirmed this works as
intended.