Try removing AT symbol for AWS
It is possible to add records with `@` on Route53 which signifies the
record applies to the root of that subdomain. When applying these
records I was not able to return any results using `dig`.

In a similar function to Google Cloud, I have removed the `@` symbol and
set it to empty.

We could remove it from the data itself, but I feel this might be less
clear.