Simplify rake tasks
There are several tasks that are only intended to be used along with
other tasks (e.g. 'validate_generate_environment' and
'local_state_check'). These make it unclear what's intended to be used
and what isn't.

This commit moves these support tasks to functions which can then be
used in the tasks they support.