Set `@` symbol if subdomain matches origin
When reviewing [this PR](https://github.com/alphagov/govuk-dns-config/pull/13) I noted
that subdomains were being set as the full origin name when using the
root domain. For example, if the origin is `example.com`, and there was
an NS record for the root domain, then it was setting the subdomain also
as `example.com`.

While technically correct due to the trailing dot, I believe it is not
as clear to read, and is inconsistent with how we have set up our
configuration for the zone already in use [ref](https://github.com/alphagov/govuk-dns-config/blob/master/publishing.service.gov.uk.yaml#L4-L7).