Move deployment config into config file
The motivation for this commit was because we are looking to migrate ~40
zones from DynDNS to using this code to deploy into GCP or AWS.

To deploy DNS, we use a Jenkins job, which also requires you to set the
Route 53 Zone ID and Google DNS and Zone names. To find this information
across each zone and set as an environment variable is unfriendly to the
user attempting to deploy the code, especially since:

a) The user (probably on 2ndline) may not have access to this
information
b) Mistakes are more likely to happen
c) These details are unlikely to change on deployments

This has not been an issue until now as we have only been using this
code to deploy to a single zone.

Hardcoding these details into each zone means it's easy to update, they
are stored in version control, and fewer mistakes will be made.