Remove validator test for whitespace in TXT records
In govuk-dns-config/412080c3 a validator test was added to reject TXT record
data containing unescaped whitespace. However, we now have a legitimate case
where we want to allow unescaped whitespace in TXT records. We want to allow the
GoogleDNS API to split on whitespace into separated quoted strings to cope with
cases where the data string is > 255 characters, e.g. DKIM or DMARC records.

For other cases people will need to escape whitespace to avoid their TXT records
being split on whitespace if this is not what they want.

Co-authored-by: James Alderman <james.alderman@digital.cabinet-office.gov.uk>