Don't append md5sums to resource titles in GCP
These exist for what appear to be historical reasons, but are causing
problems now.

To change a resource, Terraform has to delete the old one and create a
new one, which usually fails in GCP due to eventual consistency issues
in the API.

Changing an existing record is supported by Terraform and works.