Upgrade GCP terraform provider to 2.20.3
This is the latest 2.X version. The 3.X versions are not compatible with
terraform 0.11.

The 1.X to 2.X upgrade doesn't include any breaking changes to the DNS
resources which we use:

https://registry.terraform.io/providers/hashicorp/google/latest/docs/guides/version_2_upgrade

This upgrade should (in theory) allow us to use
`GOOGLE_OAUTH_ACCESS_TOKEN` to authenticate to Google, instead of giving
Jenkins long lived credentials. We'll need to update the Jenkins job to
allow that.