WIP - script fallback to secondary CDN
This is an attempt at scripting the DNS changes needed to switch from
Fastly (the primary CDN) to AWS CloudFront (the secondary CDN).

We already know how to do this with terraform, but we've discovered that
it can be difficult to use terraform while Fastly is down.

Both the terraform registry and rubygems.org sit behind Fastly. This
means that simple commands like `terraform init` and `bundle install`
can't be relied upon in a Fastly outage.

This commit attempts to script the entire process using the `aws` and
`gcloud` CLIs. It's orchestrated with ruby, but relies only on the
standard library (so no rubygems). This would mean it would be possible
to automate the DNS change with only the two CLIs as dependencies.

I think this demonstrates that it is actually quite difficult to script
this seemingly simple process. I'm not sure I'd want to rely on this
code in an incident.

That probably means we should consider other approaches, such as
providing a mirror for our terraform providers which does not depend on
Fastly (or other third-party infrastructure).